backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
asyncpg==0.29.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.2
httpx==0.27.0
pandas==2.2.1
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.1.post1
scipy==1.13.0
joblib==1.3.2
redis==5.0.3
python-dotenv==1.0.1
rapidfuzz==3.8.1
feedparser==6.0.11
google-generativeai==0.4.1
apify-client==1.6.2
soccerdata==1.9.0